+++ CHANGELOG.md
@@ -19,11 +19,11 @@
 * `-` Remove some useless graphs and stats.
 * `-` Remove access rules from Categories.
 * `-` Remove release descriptions (not shown since a while ago), kept release notes.
+* `-` Remove subscription level support; use Joomla view access levels (as you should be doing since 2013 at the latest).
 * `~` Removed jQuery dependence [gh-168]
 * `~` Replaced jqPlot with Chart.js.
 * `~` Upgrade deprecated Joomla API calls to Joomla 3.9.
 * `~` Work towards future PHP 7.4 support.
-* `~` Optimize query caching in SubscriptionIntegration.
 * `#` [HIGH] Input filter initialization incompatible with Joomla 4
 * `#` [HIGH] MySQL script error could lead to unusable backend interface
 * `#` [HIGH] XML streams broken on hosts with short PHP tags enabled